DAB+: fix bugs by using explicit signaling in AudioSpecificConfig
This circumvents a bug of FAAD2 which occurs with SBR services on systems where
the `char` data type is unsigned:

The struct `mp4AudioSpecificConfig` has a variable `sbr_present_flag` of type
char. In `AudioSpecificConfigFromBitfile` within `mp4.c` this variable is
initially set to `-1`. Unfortunately it cannot be assumed that char is signed!
On systems where char is unsigned the variable therefore gets set to `255`.
At the end of the mentioned function, implicit SBR signalling is detected by
checking the core sample rate for a specific bound - but only if
`sbr_present_flag` is still `-1`.

This check (here indirectly invoked by `NeAACDecInit2`) is therefore skipped on
systems with unsigned char like e.g. Raspbian Jessy. In this case, only the
core sample rate is returned for SBR services instead of the actual output
sample rate. To fix this problem, one can compile FAAD2 using `-fsigned-char`.
On systems providing signed char by default FAAD2 works fine.

The problem just affects the returned sample rate (as it seems), as the further
processing is able to detect the SBR extension within the audio stream.

Using fdk-aac this commit furthermore fixes the initial "interrupted" audio
with SBR services until the SBR extension is found (after about half a second).