FIC: simplify observer interface
Initially the observer interface for the FIC decoder was designed to
work with two steps:

- FIC decoder: notify observer that new data is available
- observer: retrieve the new data

The reason was that GTK requires GUI changes (that would certainly be
consequence of a FIC change) to be carried out only by the GUI thread.
Hence after the first step, the GUI thread was notified and executed the
second step.

However the FIC decoder does not have to know about this, therefore the
observer interface now only consists of the following, single step:

- FIC decoder: notify observer of new data