Add secondary component audio services
In addition to the primary component, an audio service may also contain
further secondary component audio services (e.g. the `BBC National DAB`
ensemble in the UK).

In the GTK version the components of such services are shown in the
service list with a `»` indicator (suffixed for the primary component,
prefixed for secondary ones).