DAB+: disable PAD decoder reset on sync/missing AU
On super frame sync or on invalid AU CRC, the PAD is resetted. However
if DAB is used or if exactly a complete super frame is missing, no reset
occurs (which is not an issue as the PAD CRC check stands in).

So for consistency, remove the extra PAD decoder resets and just act
according to spec if no data_stream_element() is present.